Move ParsingPDAFactory to test folder
